trackpulse/api/pyproject.toml
[tool.poetry]
name = "trackpulse-api"
version = "0.1.0"
description = ""
authors = ["Marc Lamy <marc@marclamy.com>", "Corentin Artaud <corentinartaud@gmail.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
questdb = "^2.0.3"
python-dotenv = "^1.0.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
pandas = "^2.2.3"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
fastapi = "^0.115.6"
pydantic-settings = "^2.6.1"
psycopg2-binary = "^2.9.10"
aiohttp = "^3.11.9"
rich = "^13.9.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
